feat: add per-user download subfolders
- add typed Library settings and bind Library section
- resolve username from current HTTP request in download pipeline
- apply username folder only in permanent storage mode
- keep cache behavior global
- update appsettings/env/docker-compose wiring
- split PathHelper tests into dedicated file and add user-subfolder coverage
