+++ .env.example
@@ -4,6 +4,11 @@ SUBSONIC_URL=http://localhost:4533
 # Path where downloaded songs will be stored on the host (only applies if STORAGE_MODE=Permanent)
 DOWNLOAD_PATH=./downloads
 
+# Per-user download subfolders (optional, default: false)
+# When true and STORAGE_MODE=Permanent, downloads are saved to:
+# {DOWNLOAD_PATH}/{subsonic_username}/Artist/Album/Track
+LIBRARY_USER_SUBFOLDERS=false
+
 # Music service to use: Deezer, Qobuz, or SquidWTF (default: SquidWTF)
 MUSIC_SERVICE=SquidWTF
