+++ octo-fiesta/appsettings.json
@@ -16,7 +16,8 @@
     "CacheDurationHours": 1
   },
   "Library": {
-    "DownloadPath": "./downloads"
+    "DownloadPath": "./downloads",
+    "UserSubfolders": false
   },
   "Deezer": {
     "Arl": "",
